Important correction to make sure we keep the references of Events and TemporalDatabases of the current State. Previously we add references on those of the previous State mixed with references to the current one. It might result in some overhead (which we can deal with later when we have more time) but at least the result of planning should be correct.
I also added a new time-point in the STN that matches with the earliest execution time (current time + repair time). Actions are added relativly to this time point.
On execution, they are placed backed relatvly to absolute GlobalStart.
The way this is dealt with at execution is still a bit hacky but paln repair now works.