This is a huge refactoring to support several new features (whose previous inplementation was incorrect/incomplete):  - decomposition without the need of an open goal. A seed a action now triggers the decomposition.  - correctly handles types of varaibles.
The code for constraints, events, and types/obect-constants is no much simpler.

Overview of changes:
 - there is only one IUnifiable: the object variable. Object variables are created when a new action is inserted (one per parameter).
   THey are possibly binded to those of the parent action.
 - create parameterized state variable: a state variable name and an object variable.
 - event are a parameterized state variable and and two object variables containing the values
 - Introduce a TypeManager that handles all types and oject constants in the system. A complete support for type inheritance should be supported now.

As a consequence,  events are now immutable (just references to object variables). We can take advantage of that to further simplfy the code.

Bad sides: search is _not_ more efficient, since the heuristic is still bad, it evens perform worse on some instances.
In fact here the more powerfull constraint propagations has some drawbacks on the search. However, this should be adressed at the search level.

Database splitting is not implemented. I don't want to patch it in an ugly way and will do it properly later.

Since, this is _very_ intrusive, I keep to my branch until I am sure it doesn't come with hidden bugs.
(But I have tested it quite a lot and I'm pretty confindent it is correct).