Moved problem definitions (vars, types, actions, ...) outside Planner to a new class Problem.
Goal here is to avoid referencing pthe planner for things that are not in its juridiction and to be able to have domain information in other places.
Problem supports revisions which are a list to updates to be made to a search node. THis would need a bit more polishing but one thing we want here is to support
updating any state with latest problem definition from the planner.

Furthermore, the Problem class should be able to handle most problem specific preprocessing.