New main method for planning in fape.Planning, supporting the different planners and configuration options. Look at README.md for usage.
We now have an abstract class APlanner that contains everything that was formely in fape.core.planning.Planner.
Subclassing it allows to easily define customs search policies / heuristics / domain analysis ...
Theere is currently two classes for it: Planner (that sticks to a lifted representation) and PGPlanner (that uses the relaxed planning graph).