Correct version of canBeStrictlyBefore and remove compilation failure due to missing file.
Remember that in an STN represented as a distance graph, a delay d between two time points a
and b is expressed as an edge (b, a, d) (from b to a with weight d).

Also a file Metric.java seems to be missing, Filip can you check whether you forgot to add it
and uncomment the lines in PlanCompFactory.