Rigid relations are now checked through constraint processing.
A rigid relation such as passable(Locatiion, Location):Boolean is encoded as a set of tuples of values:
values("passable") = { (l1, l2, true), (l2, l3, true), ...}
A condition as passable(L1, L2) == true; is translated as
exists l,l' in dom(L1),dom(L2) such as (l, l', true) in values("passable").

The resulting constraints are propagated every time a variable of the relation is binded (reducing the domains
of the other variables).

This already improves performance since it allows for an early propagation of reachibility constraints.
However implementation is still a bit brute force and a lot of computation can be avoided through preprocessing.