Add an option for decomposition to contain ActionConditions instead of Actions
An action condition has an action name, a set of parameters and two timepoints.
It can be fulfilled/supported by an action with the same whose parameters and
time points are equal to those of the action condition.

A boolean parameter in the constructor of a Decomposition can be set to true
to replace Actions (that would be inserted into the plan) with ActionCOnditions
(that must be fulfilled through search).