Brings more general parsing capabilities.
Main change is to consider all statements (temporal statements, temporal
constraints, binding constraints, resources statements, sub task...) in a
single class.

As a consequence, we have much more general parser, with fewer restrictions on
the syntax. The reason for that is that the distinction between the different
statements is done while building the model (and while while parsing) which
allows a much stronger inference of types.