Update to support version 0.6.1 of the ANML parser.
Changes here are minimal, since most work was done in the parser.

It is now possible:
  - to write subtasks anywhere in an action (even outside of a decomposition),
    and in the problem definition (this replaces the seed action mechanism).
  - subtasks can be temporally qualified
  - constant functions don't need a temporal qualifier anymore.
  - for predicates (functions of type boolean) it is not necessary to write
    "== true" for a persistence condition.

A new syntax for duration is available:
  - duration := 10;
  - duration :in [10, 20]; (to express uncertainty)

Replacing the integers by functions is now parsed but not supported by the
parser yet.

Most domains have been migrated to take advantage of this new syntax. Some
need further work (will be done as a big domain clean-up).