[0.5.2] Improvements on MMV and FastIDC dispatcher.
Both are faster and MMV gains an incremental version.
MMV however has some corner cases that makes it not completely usable in a
dispatcher mode (problems in building the STN in any order).
Hence FastIDC is made the default dispatcher, it gains an APSP that is used to
get bounds and check cycles.