Support for "virtual time points".
Those are time points that are not relevant for execution (they won't be
dispatched and are not contingent) and are rigidly tied to another one.

This is handled at the STNUManager level by maintaining a list of virtual timepoints, associated with a distance to another "real" timepoint.
When a constraint is added to a virtual timepoint, it is translated to a constraint on its "real" timepoint.

Classes deriving from STNUManager now share much more code for handling time points (and teir type), recording constraints and virtaul timepoints handling.