Make use of virtual time points (rigidly defined wrt another).
Those allow to have much smalled STNUs, resulting in lower CPU and memory usage.
It also alows to distinguish between time points that are just a
programming/modeling facility and those that are actually relevant for acting
(encoding a controllable choice).

This builds upon improvements in both the parser and the STNU manager.