Bug fix in getCleanState and added an applyChronicle to State.
getCleanState was performing a lookup on the new state instead of the previous (leading to crash).
appplyChronicle merges any instance of Chronicle into the state. This is primarily intented for
chronicles directly created from an anml string. (other types of chronicles alrady have custom handlers).

Also does some harmonization on int vs long: all internal time are now ints. The only remaining long
is for the current time (in milliseconds).