Added a directory 'domains/test-domains' containing simple domains that should always be solvable.
This directory should contains very simple problems that the planner should
solve with no effort. The intended objective is to be able to check that there
is no regressions (after a commit, all domains in this directory should still be solvable).

If a problem is not solvable while it should be. It should be placed in the "buggy" directory so developers can check it and move it her once the problem is solved.