When looking for supporters, checks if one is necessary.
This is the case if the end of the supporting event is exactly one time unit before the
start of the consuming event and they are necessarily on the same state variable.
In this case, this is the only applicable resolver.

This allows to avoid a lot of search when having high level actions because by defining
persistences conditions between the subtasks. Those will directly provide the causal links
between the subtasks.