Important clean up in the way temporal constraints are represented.
Constraint (binding and temporal) are now considered separately to statements (assignments, persistence an transitions).
Reasonning is done at the level of abstract actions to identify anchored timepoints and limit the number of
insertions in the STN (this was previously done on the creation of an action instance).

Duration are now directly translated into temporal constraints.